Added a 'tensor_permute_in_place' fxn for previous usecase
The fxn 'tensor_permute_in_place' can simply replace the values in the
given function.
The modifed function tensor_permute' also was made to
use this newly added function.
Also now the usecases of test cases 'baseops' and 'onstack' are
covered as they are supposed to work on the original array and
original array dimensions directly