test outputs and errors are dumped into the same file
It was getting annoying in the 'error chain' branch to monitor both
the stdout and stderr files separately, and also it was getting
difficult to track the sequences of errors and outputs.

So changed the tester code and the testing files to now dump both
errors and outputs in the same file. To be merged in another branch later